+++ src/photosynth_pmodel.mod.f90
@@ -215,7 +215,7 @@ contains
       lue = kphio * mprime * c_molmass  ! in g CO2 m-2 s-1 / (mol light m-2 s-1)
 
       ! Vcmax after accounting for Jmax limitation
-      vcmax = kphio  * ppfd * out_optchi%mjoc * mprime / out_optchi%mj
+      vcmax = kphio * ppfd * out_optchi%mjoc * mprime / out_optchi%mj ! mol m-2 s-1
 
 
     else if (method_jmaxlim == "wang17") then